Merged PR 9: Added support for RabbitMQ, a worker engine, and basic job file storage via Attachments
Among changes to the Docker scripts, this adds a whole new project to the solution. The Slicer.

The slicer takes in work items off a RabbitMQ queue, and does some as yet undermined action on them then updates the database with the results.

There is an `ISlicingService` added to the Web Application to easily facilitate this operation.

This was done in the branch without the RavenDB Models, so I had to add them to the common project.