Merged PR 34: Job IDs, User IDs, and authentication, oh my!
Job IDs are now GUIDs
User IDs are now GUIDs
There is an authentication API that can be hit to login, logout, register, and change password.

I integrated some more of the framework features in, for example each controller now takes a UserManager in order to retrieve the user properly from the database.

Also, I've remove usages of the [Authorize] attribute, as its behavior is just to render the login page to the user, instead of just returning a 401.